backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Queue
redis==5.2.1
rq==1.16.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI / LLM
anthropic==0.42.0
httpx==0.28.1

# Crypto
cryptography==44.0.0

# Observability
prometheus-client==0.21.1
structlog==24.4.0
python-json-logger==3.2.1

# Email
aiosmtplib==3.0.2

# YAML
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Utils
python-dotenv==1.0.1
